move cmd package into core directory
