+++ core/cmd/client_test.go
@@ -3,7 +3,7 @@ package cmd_test
 import (
 	"testing"
 
-	"github.com/smartcontractkit/chainlink/cmd"
+	"github.com/smartcontractkit/chainlink/core/cmd"
 	"github.com/smartcontractkit/chainlink/internal/cltest"
 	"github.com/smartcontractkit/chainlink/store/models"
 	"github.com/stretchr/testify/assert"
@@ -95,7 +95,7 @@ func TestDiskCookieStore_Retrieve(t *testing.T) {
 		wantError bool
 	}{
 		{"missing", config.RootDir(), true},
-		{"correct fixture", "../internal/fixtures", false},
+		{"correct fixture", "../../internal/fixtures", false},
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
@@ -178,7 +178,7 @@ func TestFileAPIInitializer_InitializeWithoutAPIUser(t *testing.T) {
 		file      string
 		wantError bool
 	}{
-		{"correct", "../internal/fixtures/apicredentials", false},
+		{"correct", "../../internal/fixtures/apicredentials", false},
 		{"no file", "", true},
 		{"incorrect file", "/tmp/doesnotexist", true},
 	}
@@ -262,7 +262,7 @@ func TestFileSessionRequestBuilder(t *testing.T) {
 		wantError             bool
 	}{
 		{"empty", "", "", true},
-		{"correct file", "../internal/fixtures/apicredentials", "email@test.net", false},
+		{"correct file", "../../internal/fixtures/apicredentials", "email@test.net", false},
 		{"incorrect file", "/tmp/dontexist", "", true},
 	}
