+++ core/cmd/doc.go
