+++ core/cmd/key_store_authenticator.go
