+++ core/cmd/renderer.go
